requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg[binary]==3.3.3
passlib==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
pydantic==2.10.2
python-dotenv==1.2.1
email-validator==2.3.0
httpx==0.27.2
stripe==10.8.0
alembic==1.13.2
pytest==8.3.3
